requirements.txt
accelerate==0.32.1
asgiref==3.8.1
bitsandbytes==0.42.0
certifi==2024.7.4
charset-normalizer==3.3.2
Django==4.2.14
django-cors-headers==4.4.0
djangorestframework==3.15.2
filelock==3.15.4
fsspec==2024.6.1
huggingface-hub==0.23.5
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
packaging==24.1
psutil==6.0.0
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scipy==1.13.1
sqlparse==0.5.1
sympy==1.13.0
tokenizers==0.19.1
torch==2.3.1
tqdm==4.66.4
transformers==4.42.4
typing_extensions==4.12.2
urllib3==2.2.2
